Add mesh visualization component and integrate into home page
